Supression d'un test
