--- dashboard/tests/views/test_main.py
@@ -1,79 +1,0 @@
-##############################################################################
-#
-#    OSIS stands for Open Student Information System. It's an application
-#    designed to manage the core business of higher education institutions,
-#    such as universities, faculties, institutes and professional schools.
-#    The core business involves the administration of students, teachers,
-#    courses, programs and so on.
-#
-#    Copyright (C) 2018 Université catholique de Louvain (http://www.uclouvain.be)
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
-#    the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    A copy of this license - GNU General Public License - is available
-#    at the root of the source code of this program.  If not,
-#    see http://www.gnu.org/licenses/.
-#
-##############################################################################
-import datetime
-
-from django.conf import settings
-from django.test import TestCase
-from django.urls import reverse
-
-from base.models.enums import academic_calendar_type
-from base.tests.factories.academic_calendar import AcademicCalendarFactory
-from base.tests.factories.academic_year import AcademicYearFactory
-from base.tests.factories.user import UserFactory
-
-
-class TestHome(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.user = UserFactory()
-        cls.url = reverse("home")
-
-    def setUp(self):
-        self.client.force_login(self.user)
-
-    def test_user_is_not_logged(self):
-        self.client.logout()
-        response = self.client.get(self.url)
-        self.assertRedirects(response, "/login/?next={}".format(self.url))
-
-    def test_template_used(self):
-        response = self.client.get(self.url)
-        self.assertTemplateUsed(response, "dashboard.html")
-
-    def test_with_online_application_not_opened(self):
-        response = self.client.get(self.url)
-        context = response.context
-        self.assertFalse(context["online_application_opened"])
-
-    def test_with_online_application_opened(self):
-        today = datetime.date.today()
-        current_academic_year = AcademicYearFactory(year=today.year, start_date=today - datetime.timedelta(days=5),
-                                                    end_date=today + datetime.timedelta(days=5))
-        AcademicCalendarFactory(academic_year=current_academic_year,
-                                reference=academic_calendar_type.TEACHING_CHARGE_APPLICATION)
-        response = self.client.get(self.url)
-        context = response.context
-        self.assertTrue(context["online_application_opened"])
-
-    def test_manage_courses_url(self):
-        response = self.client.get(self.url)
-        context = response.context
-        self.assertEqual(context["manage_courses_url"], settings.OSIS_MANAGE_COURSES_URL)
-
-    def test_osis_vpn_help_url(self):
-        response = self.client.get(self.url)
-        context = response.context
-        self.assertEqual(context["osis_vpn_help_url"], settings.OSIS_VPN_HELP_URL)
